feat(router): introduce virtual path configuration
This commit introduces a new feature in the router, the virtual path configuration. This feature allows users to configure virtual paths that do not exist in the Apache server but can be mapped to existing paths. The operations related to this configuration are accessible from the path `/spaship-proxy/api/v1`, and the API specification is available at `/spaship-proxy/api/v1/docs`.

A Containerfile has been added to build the image.

For more details, refer to the updated README.md.