feat(router): virtual path configuration from router (#2632)
* feat(router): introduce virtual path configuration

This commit introduces a new feature in the router, the virtual path configuration. This feature allows users to configure virtual paths that do not exist in the Apache server but can be mapped to existing paths. The operations related to this configuration are accessible from the path `/spaship-proxy/api/v1`, and the API specification is available at `/spaship-proxy/api/v1/docs`.

A Containerfile has been added to build the image.

For more details, refer to the updated README.md.

* fix: update packages to fix vulnerabilities

This commit updates the packages as per the recommendations of `npm audit`. The vulnerabilities identified in the `@spaship/router` package have been addressed by updating the respective packages to their latest versions.

* Fix: Handle 'TypeError: Cannot destructure property' issue and update README

- Resolved a json parse error issue.
- Updated the content of the README file for better clarity and understanding.

* chore: Variable renaming, API naming convention changes, and API visibility configuration

- Renamed variables to camelCase as suggested in PR
- Changed API naming convention from camelCase to lowercase with hyphen separators
- Updated OpenAPI specification and README to reflect these changes
- Added feature to configure visibility of API specifications under /docs path